requirements.txt
fastapi
uvicorn[standard]
alembic
python-dotenv
faker